+++ src/hyper_tokio/hyper_ext.rs
@@ -125,63 +125,4 @@ impl HyperExtensions {
             _ => Err(Box::new(SlackEventAbsentSignatureError::new())),
         }
     }
-
-    pub fn generate_multipart_boundary() -> String {
-        format!(
-            "----WebKitFormBoundarySlackMorphismRust{}",
-            chrono::Utc::now().timestamp()
-        )
-    }
-
-    pub fn create_multipart_file_content<'p, PT, TS>(
-        fields: &'p PT,
-        multipart_boundary: &str,
-        file_name: &str,
-        file_content_type: &str,
-        file_content: &[u8],
-    ) -> AnyStdResult<Bytes>
-    where
-        PT: std::iter::IntoIterator<Item = (&'p str, Option<TS>)> + Clone,
-        TS: AsRef<str> + 'p + Send,
-    {
-        let mut output = BytesMut::with_capacity(file_content.len() + 512);
-        output.write_str("\r\n")?;
-        output.write_str("\r\n")?;
-        output.write_str("--")?;
-        output.write_str(multipart_boundary)?;
-        output.write_str("\r\n")?;
-        output.write_str(&format!(
-            "Content-Disposition: form-data; name=\"file\"; filename=\"{}\"",
-            file_name
-        ))?;
-        output.write_str("\r\n")?;
-        output.write_str(&format!("Content-Type: {}", file_content_type))?;
-        output.write_str("\r\n")?;
-        output.write_str(&format!("Content-Length: {}", file_content.len()))?;
-        output.write_str("\r\n")?;
-        output.write_str("\r\n")?;
-        output.put_slice(file_content);
-
-        for (k, mv) in fields.clone().into_iter() {
-            if let Some(v) = mv {
-                let vs = v.as_ref();
-                output.write_str("\r\n")?;
-                output.write_str("--")?;
-                output.write_str(multipart_boundary)?;
-                output.write_str("\r\n")?;
-                output.write_str(&format!("Content-Disposition: form-data; name=\"{}\"", k))?;
-                output.write_str("\r\n")?;
-                output.write_str(&format!("Content-Length: {}", vs.len()))?;
-                output.write_str("\r\n")?;
-                output.write_str("\r\n")?;
-                output.write_str(vs)?;
-            }
-        }
-
-        output.write_str("\r\n")?;
-        output.write_str("--")?;
-        output.write_str(multipart_boundary)?;
-
-        Ok(output.freeze())
-    }
 }
