+++ src/lib.rs
@@ -101,6 +101,7 @@ mod scroller;
 pub mod signature_verifier;
 pub mod socket_mode;
 
+mod multipart_form;
 mod token;
 
 #[cfg(feature = "hyper")]
